requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
black==22.8.0
click==8.1.3
coverage==6.5.0
decorator==5.1.1
Django==4.1.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
executing==1.1.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
jsonschema==4.16.0
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
traitlets==5.4.0
uritemplate==4.1.1
wcwidth==0.2.5
